+++ server/api/orders.js
@@ -3,7 +3,7 @@ const {Order} = require('../db/models')
 module.exports = router
 
 router.get('/', (req, res, next) => {
-  Order.findAll()
+  Order.findAll({ include: [{ all: true, nested: true }]})
     .then(Orders => res.json(Orders))
     .catch(next)
 })
